+++ plugins/org.wso2.developerstudio.eclipse.esb.dashboard.templates/src/org/wso2/developerstudio/eclipse/esb/dashboard/templates/wizard/ConnectorArtifact.java
@@ -1,0 +1,83 @@
+/*
+ *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.developerstudio.eclipse.esb.dashboard.templates.wizard;
+
+public class ConnectorArtifact {
+
+    private String name;
+    private String version;
+    private String serverRole;
+    private String type;
+    private String groupId;
+
+    private String file;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    public String getServerRole() {
+        return serverRole;
+    }
+
+    public void setServerRole(String serverRole) {
+        this.serverRole = serverRole;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public void setFile(String file) {
+        this.file = file;
+    }
+
+    public String getFile() {
+        return file;
+    }
+
+    public boolean isAnonymous() {
+        return name != null ? false : true;
+    }
+
+    public void setGroupId(String groupId) {
+        this.groupId = groupId;
+    }
+
+    public String getGroupId() {
+        return groupId;
+    }
+
+}
